Updated repository to support properly Zeppelin. Changed java version from 8 to 7. Edited several configuration files of hadoop to deploy YARN.

+++ Vagrantfile
@@ -17,14 +17,17 @@ Vagrant.configure(2) do |config|
       cp /vagrant/shared/ubuntu1.pub /home/vagrant/.ssh/id_rsa.pub
       cat /vagrant/shared/ubuntu1.pub >> /home/vagrant/.ssh/authorized_keys
 
-      #if [ -f /vagrant/zeppelin-0.6.2-bin-all.tgz ]
-      #then
-      #  sudo wget http://apache.rediris.es/zeppelin/zeppelin-0.6.2/zeppelin-0.6.2-bin-all.tgz -P /vagrant/
-      #fi
-      #sudo tar -zxvf /vagrant/zeppelin-0.6.2-bin-all.tgz -C /usr/local/
-      #sudo ln -sf /usr/local/zeppelin-0.6.2-bin-all/ /usr/local/zeppelin
-      #sudo chown -R vagrant /usr/local/zeppelin
-      #sudo chown -R vagrant /usr/local/zeppelin-0.6.2-bin-all/
+      if [ ! -f /vagrant/zeppelin-0.6.2-bin-all.tgz ]
+      then
+        sudo wget http://apache.rediris.es/zeppelin/zeppelin-0.6.2/zeppelin-0.6.2-bin-all.tgz -P /vagrant/
+      fi
+      if [ -f /vagrant/zeppelin-0.6.2-bin-all.tgz ]
+      then
+      sudo tar -zxvf /vagrant/zeppelin-0.6.2-bin-all.tgz -C /usr/local/
+      sudo ln -sf /usr/local/zeppelin-0.6.2-bin-all/ /usr/local/zeppelin
+      sudo chown -R vagrant /usr/local/zeppelin
+      sudo chown -R vagrant /usr/local/zeppelin-0.6.2-bin-all/
+      fi
     SHELL
   end
 
@@ -43,7 +46,7 @@ Vagrant.configure(2) do |config|
   end
 
   config.vm.provider "virtualbox" do |vb|
-    vb.memory = 4096
+    vb.memory = 6192
     vb.cpus = 2
   end
 
@@ -58,11 +61,15 @@ Vagrant.configure(2) do |config|
     #Install outdated software
     sudo apt-get -y upgrade
 
-    #Automatically Accept Java8 License
+    #Install unzip command
+    sudo apt-get -y install unzip
+
+    #Automatically Accept Java 7 and 8 License
+    sudo echo oracle-java7-installer shared/accepted-oracle-license-v1-1 select true | sudo /usr/bin/debconf-set-selections
     sudo echo oracle-java8-installer shared/accepted-oracle-license-v1-1 select true | sudo /usr/bin/debconf-set-selections
 
-    #Install Java8
-    sudo apt-get install -y oracle-java8-installer
+    #Install Java7
+    sudo apt-get install -y oracle-java7-installer
 
     #Download Hadoop 2.7.3 Binary
     if [ ! -f "/vagrant/hadoop-2.7.3.tar.gz" ]
@@ -70,6 +77,8 @@ Vagrant.configure(2) do |config|
       echo "Downloading Hadoop 2.7.3 Binaries"
     sudo wget http://ftp.cixug.es/apache/hadoop/common/hadoop-2.7.3/hadoop-2.7.3.tar.gz -P /vagrant/
     fi
+    if [ -f "/vagrant/hadoop-2.7.3.tar.gz" ]
+    then
     sudo tar -zxvf /vagrant/hadoop-2.7.3.tar.gz -C /usr/local/
     sudo ln -sf /usr/local/hadoop-2.7.3/ /usr/local/hadoop
     sudo chown -R vagrant /usr/local/hadoop/
@@ -77,6 +86,7 @@ Vagrant.configure(2) do |config|
     sudo tar -zxvf /vagrant/shared/hadoop-files.tar.gz -C /usr/local/hadoop/etc/
     sudo cp /usr/local/hadoop/etc/hadoop-files/* /usr/local/hadoop/etc/hadoop/
     sudo rm -rf /usr/local/hadoop/etc/hadoop-files
+    fi
 
     #Download Spark 1.6.2
     if [ ! -f "/vagrant/spark-1.6.2-bin-hadoop2.6.tgz" ]
@@ -84,24 +94,32 @@ Vagrant.configure(2) do |config|
       echo "Downloading Spark 1.6.2 Binaries"
       sudo wget http://d3kbcqa49mib13.cloudfront.net/spark-1.6.2-bin-hadoop2.6.tgz -P /vagrant/
     fi
+    if [ -f "/vagrant/spark-1.6.2-bin-hadoop2.6.tgz" ]
+    then
     sudo tar -zxvf /vagrant/spark-1.6.2-bin-hadoop2.6.tgz -C /usr/local/
     sudo ln -sf /usr/local/spark-1.6.2-bin-hadoop2.6/ /usr/local/spark
     sudo chown -R vagrant /usr/local/spark/
     sudo chown -R vagrant /usr/local/spark-1.6.2-bin-hadoop2.6
+    fi
 
     #Setup bashrc
-    echo "export JAVA_HOME=/usr/lib/jvm/java-8-oracle/" >> /home/vagrant/.bashrc
+    echo "export JAVA_HOME=/usr/lib/jvm/java-7-oracle/" >> /home/vagrant/.bashrc
     echo "export HADOOP_HOME=/usr/local/hadoop" >> /home/vagrant/.bashrc
-    echo "export HADOOP_CONF_DIR=/usr/local/hadoop/etc/hadoop"
+    echo "export HADOOP_CONF_DIR=/usr/local/hadoop/etc/hadoop" >> /home/vagrant/.bashrc
     echo "export SPARK_HOME=/usr/local/spark" >> /home/vagrant/.bashrc
     echo "export PATH=$PATH:/usr/local/hadoop/bin:/usr/local/spark/bin" >> /home/vagrant/.bashrc
 
+    #Disable IPV6
+    sudo echo "net.ipv6.conf.all.disable_ipv6 = 1" >> /etc/sysctl.conf
+    sudo echo "net.ipv6.conf.default.disable_ipv6 = 1" >> /etc/sysctl.conf
+    sudo echo "net.ipv6.conf.lo.disable_ipv6 = 1" >> /etc/sysctl.conf
+
+    sudo sysctl -p
+
     #Configure hostname resolution
+    sudo echo "127.0.0.1 localhost" > /etc/hosts
     sudo echo "10.0.0.10 ubuntu1.tallerhadoop1.org ubuntu1" >> /etc/hosts
     sudo echo "10.0.0.11 ubuntu2.tallerhadoop1.org ubuntu2" >> /etc/hosts
-    cat /etc/hosts | grep -v ^127.0.1.1 | grep -v ^f | grep -v ^: | grep -v ^# > hosts
-    sudo cp hosts /etc/hosts
-    rm hosts
 
   SHELL
 end
